Update IPAM pool monitor's pool state logic and language (#1126)
* move and rename ipstate and change allocated to assigned


typo

* build ip pool state and populate struct pre-reconcile


* upcast to int64 and move scaler fields to metastate


* delint


* fix tests


* upcast usage of int to int64 in monitor fake tests


* update remaining usages of allocate to assign

+++ cns/fakes/monitor.go
@@ -11,7 +11,7 @@ import (
 )
 
 type MonitorFake struct {
-	IPsNotInUseCount  int
+	IPsNotInUseCount  int64
 	NodeNetworkConfig *v1alpha.NodeNetworkConfig
 }
 
@@ -29,8 +29,8 @@ func (*MonitorFake) Reconcile() error {
 
 func (f *MonitorFake) GetStateSnapshot() cns.IpamPoolMonitorStateSnapshot {
 	return cns.IpamPoolMonitorStateSnapshot{
-		MaximumFreeIps:           int(float64(f.NodeNetworkConfig.Status.Scaler.BatchSize) * (float64(f.NodeNetworkConfig.Status.Scaler.ReleaseThresholdPercent) / 100)), //nolint:gomnd // it's a percent
-		MinimumFreeIps:           int(float64(f.NodeNetworkConfig.Status.Scaler.BatchSize) * (float64(f.NodeNetworkConfig.Status.Scaler.RequestThresholdPercent) / 100)), //nolint:gomnd // it's a percent
+		MaximumFreeIps:           int64(float64(f.NodeNetworkConfig.Status.Scaler.BatchSize) * (float64(f.NodeNetworkConfig.Status.Scaler.ReleaseThresholdPercent) / 100)), //nolint:gomnd // it's a percent
+		MinimumFreeIps:           int64(float64(f.NodeNetworkConfig.Status.Scaler.BatchSize) * (float64(f.NodeNetworkConfig.Status.Scaler.RequestThresholdPercent) / 100)), //nolint:gomnd // it's a percent
 		UpdatingIpsNotInUseCount: f.IPsNotInUseCount,
 		CachedNNC:                *f.NodeNetworkConfig,
 	}
